refactor: rearrange operation result analysis
