--- src/Lando/LowLevel/Enums/OperationResultType.cs
@@ -1,16 +1,0 @@
-namespace Lando.LowLevel.Enums
-{
-	public enum OperationResultType
-	{
-		Success,
-		Failed,
-		CardRemoved,
-		NoReadersAvailable,
-		InvalidHandle,
-		ReaderUnavailable,
-		NoCard,
-		UnknownReader,
-		CardResetted,
-		SystemError
-	}
-}
